+++ Automate/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Added [full support for Junimo chests](README.md#junimo-chests), which enables global supply chains.
 * Removed old `config.json` migrations. If you changed the mod options in Automate 1.17.3 or earlier, and haven't played since then, you may need to reconfigure them.
+* The default machine priorities are now set in `assets/data.json`. You can override an entry in your `config.json`.
 
 ## 1.19.1
 Released 04 January 2021.